+++ packages/truffle-codec/lib/decode/abi.ts
@@ -8,7 +8,8 @@ import decodeValue from "./value";
 import { AbiDataPointer, DataPointer } from "../types/pointer";
 import { AbiMemberAllocation } from "../types/allocation";
 import { abiSizeForType, isTypeDynamic } from "../allocate/abi";
-import { EvmInfo, DecoderOptions } from "../types/evm";
+import { EvmInfo } from "../types/evm";
+import { DecoderOptions } from "../types/options";
 import { DecoderRequest, GeneratorJunk } from "../types/request";
 import { StopDecodingError } from "../types/errors";
