+++ packages/truffle-codec/lib/decode/event.ts
@@ -5,7 +5,8 @@ import decodeValue from "./value";
 import read from "../read";
 import { Types, Values, Conversion as ConversionUtils } from "truffle-codec-utils";
 import { EventTopicPointer } from "../types/pointer";
-import { EvmInfo, DecoderOptions } from "../types/evm";
+import { EvmInfo } from "../types/evm";
+import { DecoderOptions } from "../types/options";
 import { DecoderRequest, GeneratorJunk } from "../types/request";
 import { StopDecodingError } from "../types/errors";
